+++ lib/src/data_sources/localization_data_source.dart
@@ -1,5 +1,5 @@
 import 'dart:ui';
 
 abstract class LocalizationDataSource {
-  Future<Map<String, dynamic>> load(Locale locale);
+  Future<Map<Locale, Map<String, dynamic>>> load(List<Locale> locales);
 }