+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     - due to how asset loading works, all languages that were declared as fallback need to be loaded beforehand.
 - BREAKING: `AssetBundleLocalizationDataSource.bundlePath` is now required.
     - if you were using the default, just pass `localizations` as the argument.
+- BREAKING: `LocalizationDataSource.load` now receives a list of locales, instead of a single locale.
+    - `AssetBundleLocalizationDataSource` now also tries to load files concurrently.
 
 # [0.8.0]